chore(dependencies): migrate to community router package
this commit updates the starter template to use the community driven
version of the stencil router instead of the one that is no longer
developed by the stencil team. in doing so, we uninstall the one under
the 'stencil' scope and install the one under the 'stencil-community'
scope.

`@stencil/router@1.0.1` and `@stencil-community/router@1.0.2` are nearly
identical in terms of contents. `@stencil/router@1.0.1` was transferred
to the community model as `@stencil-community/router@1.0.1` (and are
entirely identical). however, v1.0.1 was never published under the
`@stencil-community` scope due to limitations of the router's publishin
infrastructure. `@stencil-community/router@1.0.2` was published with
minor changes to the original router necessary to publish the package.
those changes can be viewed at https://github.com/stencil-community/stencil-router/compare/v1.0.1...v1.0.2